removing deprecated test files and other files no longer needed after switching from json to direct to postgres
