--- resources/test_files/parser_testing/multiple_html_files/test_1.html
@@ -1,1 +1,0 @@
-test
