--- resources/test_files/parser_testing/multiple_html_files/test_2.html
@@ -1,1 +1,0 @@
-test
