--- resources/test_files/parser_testing/test_2.html
@@ -1,1 +1,0 @@
-test
