+++ src/__init__.py
@@ -1,5 +1,3 @@
-from . import cleaner
 from . import parser
 from . import scraper
 from . import tools
-from . import updater
